backend/django_api/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
mlflow==2.13.0
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.4.2
statsmodels==0.14.2
joblib==1.4.2
requests==2.32.3
tensorflow==2.16.1
gunicorn==22.0.0
